+++ src/app/pdf-viewer/pdf-viewer.component.spec.ts
@@ -1,8 +1,11 @@
-import { TestBed, async, ComponentFixture } from '@angular/core/testing';
+import { TestBed, ComponentFixture, waitForAsync } from '@angular/core/testing';
 import { Component } from '@angular/core';
 
 import { PdfViewerComponent } from './pdf-viewer.component';
 import { PdfViewerModule } from './pdf-viewer.module';
+
+import { GlobalWorkerOptions } from 'pdfjs-dist';
+import * as PDFJS from 'pdfjs-dist/build/pdf';
 
 @Component({
   template: `
@@ -24,7 +27,7 @@ describe('AppComponent', () => {
     };
   }
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [TestComponent],
       imports: [PdfViewerModule]
@@ -110,7 +113,9 @@ describe('AppComponent', () => {
       expect((pdfViewer as any).getDocumentParams()).toEqual({
         url: src,
         cMapUrl,
-        cMapPacked: true
+        cMapPacked: true,
+        enableXfa: true,
+        isEvalSupported: false
       });
     });
 
@@ -121,7 +126,9 @@ describe('AppComponent', () => {
       expect((pdfViewer as any).getDocumentParams()).toEqual({
         url: src,
         cMapUrl,
-        cMapPacked: true
+        cMapPacked: true,
+        enableXfa: true,
+        isEvalSupported: false
       });
     });
 
@@ -133,8 +140,57 @@ describe('AppComponent', () => {
       expect((pdfViewer as any).getDocumentParams()).toEqual({
         url: srcUrl,
         cMapUrl,
-        cMapPacked: true
+        cMapPacked: true,
+        enableXfa: true,
+        isEvalSupported: false
       });
     });
   });
+
+  describe('pdf.worker location', () => {
+    const curPdfJsVersion = (PDFJS as any).version;
+
+    beforeEach(() => {
+      (window as any).pdfWorkerSrc = undefined;
+      (window as any)["pdfWorkerSrc1.2.3"] = undefined;
+      (window as any)[`pdfWorkerSrc${curPdfJsVersion}`] = undefined;
+
+    });
+
+    it('should default to the cdn', () => {
+      pdfViewerFixture = TestBed.createComponent(PdfViewerComponent);
+      pdfViewer = pdfViewerFixture.debugElement.componentInstance;
+
+      expect(GlobalWorkerOptions.workerSrc).toBe(`https://cdn.jsdelivr.net/npm/pdfjs-dist@${curPdfJsVersion
+        }/legacy/build/pdf.worker.min.js`);
+    })
+
+    it('should support global override', () => {
+      (window as any).pdfWorkerSrc = 'globaloverride';
+
+      pdfViewerFixture = TestBed.createComponent(PdfViewerComponent);
+      pdfViewer = pdfViewerFixture.debugElement.componentInstance;
+
+      expect(GlobalWorkerOptions.workerSrc).toBe('globaloverride');
+    })
+
+    it('should default to the cdn when version override does not match version', () => {
+      (window as any)["pdfWorkerSrc1.2.3"] = 'globaloverride';
+
+      pdfViewerFixture = TestBed.createComponent(PdfViewerComponent);
+      pdfViewer = pdfViewerFixture.debugElement.componentInstance;
+
+      expect(GlobalWorkerOptions.workerSrc).toBe(`https://cdn.jsdelivr.net/npm/pdfjs-dist@${curPdfJsVersion
+        }/legacy/build/pdf.worker.min.js`);
+    })
+
+    it('should take version override with version match', () => {
+      (window as any)[`pdfWorkerSrc${curPdfJsVersion}`] = 'globaloverride';
+
+      pdfViewerFixture = TestBed.createComponent(PdfViewerComponent);
+      pdfViewer = pdfViewerFixture.debugElement.componentInstance;
+
+      expect(GlobalWorkerOptions.workerSrc).toBe(`globaloverride`);
+    })
+  })
 });